+++ contrib/scripts/transactions.sh
@@ -14,6 +14,9 @@ set -e
 echo "Running transaction tests."
 
 contrib=$GOPATH/src/github.com/dgraph-io/dgraph/contrib
+
+go test -v $contrib/integration/testtxn/main_test.go
+
 source $contrib/scripts/functions.sh
 
 rm -rf $BUILD/p* $BUILD/w*