+++ dgraph/cmd/server/http_test.go
@@ -124,7 +124,7 @@ func commitWithTs(keys []string, ts uint64) error {
 }
 
 func TestTransactionBasic(t *testing.T) {
-	require.NoError(t, alterSchema(`{"drop_all": true}`))
+	require.NoError(t, dropAll())
 	require.NoError(t, alterSchema(`name: string @index(term) .`))
 
 	q1 := `